+++ providers/openfeature-provider-flagd/tests/e2e/file/conftest.py
@@ -14,6 +14,8 @@ feature_list = {
     "~grace",
     "~contextEnrichment",
     "~deprecated",
+    "~fractional-v1",
+    "~operator-errors",
 }