+++ providers/openfeature-provider-flagd/tests/e2e/rpc/conftest.py
@@ -10,7 +10,8 @@ feature_list = [
     "~sync",
     "~metadata",
     "~deprecated",
-    "~forbidden",
+    "~fractional-v1",
+    "~operator-errors",
 ]